The BlockPI and BlockPID integral saturation meaning changed
It used to be KI* integral(error), it is now integral(KI*error)
because this allows the integral block to saturate the integral
with scaled value so that the maximum integrator effect is
indepdent of gain. It used to be really annoying because the
maximum integrator effect would constantly change with gain.